Chore: 프로젝트 도메인 계층 구조 디렉토리 설정
