--- src/main/java/com/chungang/capstone/openstep/apiPayload/code/BaseCode.java
@@ -1,6 +1,0 @@
-package com.chungang.capstone.openstep.apiPayload.code;
-
-public interface BaseCode {
-    public ReasonDTO getReason();
-    public ReasonDTO getReasonHttpStatus();
-}
