--- src/main/java/com/chungang/capstone/openstep/apiPayload/code/BaseErrorCode.java
@@ -1,6 +1,0 @@
-package com.chungang.capstone.openstep.apiPayload.code;
-
-public interface BaseErrorCode {
-    public ErrorReasonDTO getReason();
-    public ErrorReasonDTO getReasonHttpStatus();
-}
