--- src/main/java/com/chungang/capstone/openstep/apiPayload/code/ReasonDTO.java
@@ -1,17 +1,0 @@
-package com.chungang.capstone.openstep.apiPayload.code;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-public class ReasonDTO {
-    private HttpStatus httpStatus;
-
-    private final boolean isSuccess;
-    private final String code;
-    private final String message;
-
-    public boolean getIsSuccess(){return isSuccess;}
-}
