--- src/main/java/com/chungang/capstone/openstep/apiPayload/exception/GeneralException.java
@@ -1,18 +1,0 @@
-package com.chungang.capstone.openstep.apiPayload.exception;
-
-import com.chungang.capstone.openstep.apiPayload.code.BaseErrorCode;
-import com.chungang.capstone.openstep.apiPayload.code.ErrorReasonDTO;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class GeneralException extends RuntimeException {
-    private BaseErrorCode code;
-    public ErrorReasonDTO getErrorReason() {
-        return this.code.getReason();
-    }
-    public ErrorReasonDTO getErrorReasonHttpStatus(){
-        return this.code.getReasonHttpStatus();
-    }
-}
