--- src/main/java/com/chungang/capstone/openstep/apiPayload/exception/handler/MemberException.java
@@ -1,10 +1,0 @@
-package com.chungang.capstone.openstep.apiPayload.exception.handler;
-
-import com.chungang.capstone.openstep.apiPayload.code.BaseErrorCode;
-import com.chungang.capstone.openstep.apiPayload.exception.GeneralException;
-
-public class MemberException extends GeneralException {
-    public MemberException(BaseErrorCode code) {
-        super(code);
-    }
-}
