+++ src/main/java/com/chungang/capstone/openstep/domain/Issue/controller/IssueController.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Issue.controller;
+
+public class IssueController {
+}
