+++ src/main/java/com/chungang/capstone/openstep/domain/Issue/converter/IssueConverter.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Issue.converter;
+
+public class IssueConverter {
+}
