+++ src/main/java/com/chungang/capstone/openstep/domain/Issue/dto/IssueRequestDTO.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Issue.dto;
+
+public class IssueRequestDTO {
+}
