+++ src/main/java/com/chungang/capstone/openstep/domain/Issue/entity/Issue.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Issue.entity;
+
+public class Issue {
+}
