+++ src/main/java/com/chungang/capstone/openstep/domain/Issue/repository/IssueRepository.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Issue.repository;
+
+public interface IssueRepository {
+}
