+++ src/main/java/com/chungang/capstone/openstep/domain/Issue/service/IssueCommandService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Issue.service;
+
+public class IssueCommandService {
+}
