+++ src/main/java/com/chungang/capstone/openstep/domain/Issue/service/IssueQueryService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Issue.service;
+
+public class IssueQueryService {
+}
