+++ src/main/java/com/chungang/capstone/openstep/domain/Member/controller/MemberController.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Member.controller;
+
+public class MemberController {
+}
