+++ src/main/java/com/chungang/capstone/openstep/domain/Member/converter/MemberConverter.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Member.converter;
+
+public class MemberConverter {
+}
