+++ src/main/java/com/chungang/capstone/openstep/domain/Member/dto/MemberRequestDTO.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Member.dto;
+
+public class MemberRequestDTO {
+}
