+++ src/main/java/com/chungang/capstone/openstep/domain/Member/entity/Member.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Member.entity;
+
+public class Member {
+}
