+++ src/main/java/com/chungang/capstone/openstep/domain/Member/repository/MemberRepository.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Member.repository;
+
+public interface MemberRepository {
+}
