+++ src/main/java/com/chungang/capstone/openstep/domain/Member/service/MemberCommandService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Member.service;
+
+public class MemberCommandService {
+}
