+++ src/main/java/com/chungang/capstone/openstep/domain/Member/service/MemberQueryService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Member.service;
+
+public class MemberQueryService {
+}
