+++ src/main/java/com/chungang/capstone/openstep/domain/Rank/controller/RankController.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Rank.controller;
+
+public class RankController {
+}
