+++ src/main/java/com/chungang/capstone/openstep/domain/Rank/converter/RankConverter.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Rank.converter;
+
+public class RankConverter {
+}
