+++ src/main/java/com/chungang/capstone/openstep/domain/Rank/dto/RankRequestDTO.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Rank.dto;
+
+public class RankRequestDTO {
+}
