+++ src/main/java/com/chungang/capstone/openstep/domain/Rank/entity/Rank.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Rank.entity;
+
+public class Rank {
+}
