+++ src/main/java/com/chungang/capstone/openstep/domain/Rank/repository/RankRepository.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Rank.repository;
+
+public interface RankRepository {
+}
