+++ src/main/java/com/chungang/capstone/openstep/domain/Rank/service/RankQueryService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Rank.service;
+
+public class RankQueryService {
+}
