+++ src/main/java/com/chungang/capstone/openstep/domain/Repo/controller/RepoController.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Repo.controller;
+
+public class RepoController {
+}
