+++ src/main/java/com/chungang/capstone/openstep/domain/Repo/dto/RepoRequestDTO.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Repo.dto;
+
+public class RepoRequestDTO {
+}
