+++ src/main/java/com/chungang/capstone/openstep/domain/Repo/dto/RespResponseDTO.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Repo.dto;
+
+public class RespResponseDTO {
+}
