+++ src/main/java/com/chungang/capstone/openstep/domain/Repo/entity/Repo.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Repo.entity;
+
+public class Repo {
+}
