+++ src/main/java/com/chungang/capstone/openstep/domain/Repo/repository/RepoRepository.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Repo.repository;
+
+public interface RepoRepository {
+}
