+++ src/main/java/com/chungang/capstone/openstep/domain/Repo/service/RepoQueryService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Repo.service;
+
+public class RepoQueryService {
+}
