+++ src/main/java/com/chungang/capstone/openstep/domain/Task/controller/TaskController.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Task.controller;
+
+public class TaskController {
+}
