+++ src/main/java/com/chungang/capstone/openstep/domain/Task/converter/TaskConverter.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Task.converter;
+
+public class TaskConverter {
+}
