+++ src/main/java/com/chungang/capstone/openstep/domain/Task/dto/TaskResponseDTO.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Task.dto;
+
+public class TaskResponseDTO {
+}
