+++ src/main/java/com/chungang/capstone/openstep/domain/Task/entity/Task.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Task.entity;
+
+public class Task {
+}
