+++ src/main/java/com/chungang/capstone/openstep/domain/Task/repository/TaskRepository.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Task.repository;
+
+public interface TaskRepository {
+}
