+++ src/main/java/com/chungang/capstone/openstep/domain/Task/service/TaskCommandService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Task.service;
+
+public class TaskCommandService {
+}
