+++ src/main/java/com/chungang/capstone/openstep/domain/Task/service/TaskQueryService.java
@@ -1,0 +1,4 @@
+package com.chungang.capstone.openstep.domain.Task.service;
+
+public class TaskQueryService {
+}
