+++ src/main/java/com/chungang/capstone/openstep/global/apiPayload/code/BaseCode.java
@@ -1,0 +1,6 @@
+package com.chungang.capstone.openstep.global.apiPayload.code;
+
+public interface BaseCode {
+    public ReasonDTO getReason();
+    public ReasonDTO getReasonHttpStatus();
+}
