+++ src/main/java/com/chungang/capstone/openstep/global/apiPayload/code/BaseErrorCode.java
@@ -1,0 +1,6 @@
+package com.chungang.capstone.openstep.global.apiPayload.code;
+
+public interface BaseErrorCode {
+    public ErrorReasonDTO getReason();
+    public ErrorReasonDTO getReasonHttpStatus();
+}
