+++ src/main/java/com/chungang/capstone/openstep/global/apiPayload/code/ReasonDTO.java
@@ -1,0 +1,17 @@
+package com.chungang.capstone.openstep.global.apiPayload.code;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ReasonDTO {
+    private HttpStatus httpStatus;
+
+    private final boolean isSuccess;
+    private final String code;
+    private final String message;
+
+    public boolean getIsSuccess(){return isSuccess;}
+}
