+++ src/main/java/com/chungang/capstone/openstep/global/apiPayload/code/status/SuccessStatus.java
@@ -1,0 +1,50 @@
+package com.chungang.capstone.openstep.global.apiPayload.code.status;
+
+import com.chungang.capstone.openstep.global.apiPayload.code.BaseCode;
+import com.chungang.capstone.openstep.global.apiPayload.code.ReasonDTO;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum SuccessStatus implements BaseCode {
+
+    // 보안
+    USER_EMAIL_LOGIN_OK(HttpStatus.OK, "AUTH2001", "회원 이메일 로그인이 완료되었습니다."),
+    USER_KAKAO_LOGIN_OK(HttpStatus.OK, "AUTH2001", "회원 카카오 로그인이 완료되었습니다."),
+    USER_KAKAO_LOGOUT_OK(HttpStatus.OK, "AUTH2002", "회원 카카오 로그아웃이 완료되었습니다."),
+    USER_NAVER_LOGIN_OK(HttpStatus.OK, "AUTH2003", "회원 네이버 로그인이 완료되었습니다."),
+    USER_DELETE_OK(HttpStatus.OK, "AUTH2004", "회원 탈퇴가 완료되었습니다."),
+    USER_REFRESH_OK(HttpStatus.OK, "AUTH2005", "토큰 재발급이 완료되었습니다."),
+    USER_REGISTER_OK(HttpStatus.OK, "AUTH2000", "회원 가입이 완료되었습니다."),
+
+    // 유저 관련 응답
+    MEMBER_OK(HttpStatus.OK, "MEMBER_1000", "성공입니다.");
+
+
+
+    private final HttpStatus httpStatus;
+    private final String code;
+    private final String message;
+
+    @Override
+    public ReasonDTO getReason() {
+        return ReasonDTO.builder()
+                .message(message)
+                .code(code)
+                .isSuccess(true)
+                .build();
+    }
+
+    @Override
+    public ReasonDTO getReasonHttpStatus() {
+        return ReasonDTO.builder()
+                .message(message)
+                .code(code)
+                .isSuccess(true)
+                .httpStatus(httpStatus)
+                .build()
+                ;
+    }
+}
