+++ src/main/java/com/chungang/capstone/openstep/global/apiPayload/exception/GeneralException.java
@@ -1,0 +1,18 @@
+package com.chungang.capstone.openstep.global.apiPayload.exception;
+
+import com.chungang.capstone.openstep.global.apiPayload.code.BaseErrorCode;
+import com.chungang.capstone.openstep.global.apiPayload.code.ErrorReasonDTO;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class GeneralException extends RuntimeException {
+    private BaseErrorCode code;
+    public ErrorReasonDTO getErrorReason() {
+        return this.code.getReason();
+    }
+    public ErrorReasonDTO getErrorReasonHttpStatus(){
+        return this.code.getReasonHttpStatus();
+    }
+}
