+++ src/main/java/com/chungang/capstone/openstep/global/apiPayload/exception/handler/MemberException.java
@@ -1,0 +1,10 @@
+package com.chungang.capstone.openstep.global.apiPayload.exception.handler;
+
+import com.chungang.capstone.openstep.global.apiPayload.code.BaseErrorCode;
+import com.chungang.capstone.openstep.global.apiPayload.exception.GeneralException;
+
+public class MemberException extends GeneralException {
+    public MemberException(BaseErrorCode code) {
+        super(code);
+    }
+}
